add bunyan and express-mongo-sanitize

+++ app/controllers/sauce.js
@@ -8,9 +8,9 @@ exports.readSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id }) // find the sauce
     .then((sauce) => {
       sauce.imageUrl = `${req.protocol}://${req.get("host")}/` + sauce.imageUrl; // add the image url
-      res.status(200).json(sauce, hateoasLinks(req, req.params.id)); // ok
-    }) 
-    .catch((error) => res.status(404).json({ error: "Sauce not found" })); // not found
+      res.status(200).json(sauce, hateoasLinks(req, sauce._id)); // ok
+    })
+    .catch((error) => res.status(404).json({ error })); // not found
 };
 
 /*****************************************************************
@@ -26,8 +26,8 @@ exports.readAllSauces = (req, res, next) => {
         return { ...sauce._doc, hateoasLink };
       });
       res.status(200).json(sauces, hateoasLinks(req, sauces._id)); // ok
-    }) 
-    .catch((error) => res.status(404).json({ error })); // not found
+    })
+    .catch((error) => res.status(400).json({ error })); // not found
 };
 
 /*****************************************************************
@@ -39,13 +39,14 @@ exports.createSauce = (req, res, next) => {
   const sauce = new Sauce({
     ...sauceObject, // add the sauce object
     imageUrl: `images/${req.file.filename}`,
-    userId : req.auth.userID,
+    userId: req.auth.userID,
   }); // create a new sauce
   sauce
     .save() // save the sauce
     .then(() => res.status(201).json(sauce, hateoasLinks(req, sauce._id))) // created
-    .catch((error) => res.status(403).json({ error })); // forbidden
+    .catch((error) => res.status(400).json({ error })); // bad request
 };
+
 
 /*****************************************************************
  *****************  UPDATE ELEMENT IN  SAUCE    ******************
@@ -59,6 +60,9 @@ exports.updateSauce = (req, res, next) => {
     : { ...req.body }; // if there is a file, add the image url
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
+      if (sauce.userId !== req.auth.userID) {
+        return res.status(403).json({ error: "You can't update this sauce" }); // forbidden
+      }
       try {
         if (sauceObject.imageUrl) {
           fs.unlinkSync(sauce.imageUrl);
@@ -66,11 +70,13 @@ exports.updateSauce = (req, res, next) => {
       } catch (error) {
         console.log(error);
       }
+      Sauce.updateOne({ _id: req.params.id }, sauceObject, { new: true })
+        .then((sauce) =>
+          res.status(200).json(sauce, hateoasLinks(req, sauce._id))
+        ) // ok
+        .catch((error) => res.status(400).json({ error })); // bad request
     })
-    .catch((error) => res.status(403).json({ error: error })); // forbidden
-  Sauce.findOneAndUpdate({ _id: req.params.id }, sauceObject, { new: true }) 
-    .then((sauce) => res.status(201).json(sauce, hateoasLinks(req, sauce._id))) // created
-    .catch((error) => res.status(500).json({ error: error })); // Internal Server Error
+    .catch((error) => res.status(404).json({ error })); // not found
 };
 
 /*****************************************************************
@@ -79,15 +85,17 @@ exports.updateSauce = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id }) // find the sauce
     .then((sauce) => {
-      fs.unlink(sauce.imageUrl , () => {
+      if (sauce.userId !== req.auth.userID) {
+        return res.status(403).json({ error: "You can't delete this sauce" }); // forbidden
+      }
+      fs.unlink(sauce.imageUrl, () => {
         Sauce.deleteOne({ _id: req.params.id })
           .then(() => res.status(204).send()) // no content
-          .catch((error) => res.status(403).json({ error })); // forbidden
+          .catch((error) => res.status(400).json({ error })); // bad request
       });
     })
-    .catch((error) => res.status(500).json({ error })); // Internal Server Error 
+    .catch((error) => res.status(400).json({ error })); // bad request
 };
-
 
 /*****************************************************************
  *****************  LIKE OR DISLIKE A SAUCE    *******************
@@ -117,13 +125,13 @@ exports.likeSauce = (req, res, next) => {
               new: true,
             })
               .then((newSauce) => {
-                res.status(201).json(newSauce, hateoasLinks(req, sauce._id)); // created
+                res.status(200).json(newSauce, hateoasLinks(req, sauce._id)); // ok
               })
               .catch((error) => {
-                res.status(403).json({ message: error }); // forbidden
+                res.status(400).json({ error }); // bad request
               });
           } else {
-            res.status(204).json({ message: "User already dislike the sauce" }); // no content
+            res.status(200).json({ message: "User already dislike the sauce" }); // ok
           }
 
           break;
@@ -139,10 +147,10 @@ exports.likeSauce = (req, res, next) => {
               { new: true }
             )
               .then((newSauce) => {
-                res.status(201).json(newSauce, hateoasLinks(req, sauce._id)); // created
+                res.status(200).json(newSauce, hateoasLinks(req, sauce._id)); // created
               })
               .catch((error) => {
-                res.status(403).json({ message: error }); // forbidden
+                res.status(400).json({ error }); // forbidden
               });
           }
           if (sauce["usersDisliked"].includes(req.auth.userID)) {
@@ -157,11 +165,11 @@ exports.likeSauce = (req, res, next) => {
             )
               .then((newSauce) => {
                 return res
-                  .status(201) // created
+                  .status(200) // ok
                   .json(newSauce, hateoasLinks(req, sauce._id));
               })
               .catch((error) => {
-                return res.status(400).json({ message: error });
+                return res.status(400).json({ error });
               });
           }
           if (
@@ -169,7 +177,7 @@ exports.likeSauce = (req, res, next) => {
             !sauce["usersLiked"].includes(req.auth.userID)
           ) {
             res
-              .status(204) // no content
+              .status(200) // no content
               .json({ message: "Don't need to dislike or undislike" });
           }
           break;
@@ -190,17 +198,17 @@ exports.likeSauce = (req, res, next) => {
               new: true,
             })
               .then((newSauce) => {
-                res.status(201).json(newSauce, hateoasLinks(req, sauce._id)); // created
+                res.status(200).json(newSauce, hateoasLinks(req, sauce._id)); // created
               })
               .catch((error) => {
-                res.status(403).json({ message: error }); // forbidden
+                res.status(400).json({ error }); // forbidden
               });
           } else {
-            res.status(204).json({ message: "User already like the sauce" }); // no content
+            res.status(200).json({ message: "User already like the sauce" }); // no content
           }
           break;
         default:
-          res.status(404).json({ message: "Bad request" }); // not found
+          res.status(400).json({ message: "Bad request" }); // bad request
       }
     });
   } catch (error) {
@@ -235,54 +243,55 @@ exports.reportSauce = (req, res, next) => {
         res.status(404).json({ error: "No sauce to report" });
       }
     })
-    .catch((error) => res.status(500).json({ error }, console.log(error)));
+    .catch((error) => res.status(404).json({ error }));
 };
 
 // hateoas links
 const hateoasLinks = (req, id) => {
-  const URI = req.protocol + "://" + req.get("host");
+  const URI = `${req.protocol + "://" + req.get("host") + "/api/sauces/"}` ;
   return [
     {
       rel: "readOne",
       method: "GET",
-      href: URI + "/api/sauces/" + id,
+      href: URI + id,
       title: "Read sauce",
     },
     {
       rel: "readAll",
       method: "GET",
-      href: URI + "/api/sauces",
+      href: URI,
       title: "Read all sauces",
     },
     {
       rel: "create",
       method: "POST",
-      href: URI + "/api/sauces",
+      href: URI ,
       title: "Create a new sauce",
     },
     {
       rel: "update",
       method: "PUT",
-      href: URI + "/api/sauces/" + id,
+      href: URI + id,
       title: "Update a sauce",
     },
     {
       rel: "delete",
       method: "DELETE",
-      href: URI + "/api/sauces/" + id,
+      href: URI + id,
       title: "Delete a sauce",
     },
     {
       rel: "likeDislike",
       method: "POST",
-      href: URI + "/api/sauces/" + id + "/like",
+      href: URI + id + "/like",
       title: "Add like or dislike",
     },
     {
       rel: "report",
       title: "Report",
-      href: URI + "/api/sauces/" + id + "/report",
+      href: URI + id + "/report",
       method: "POST",
     },
   ];
 };
+
